Change behavior of json parser
The /Target/materials can be either an object or an object array. In the 1st case the target is composed of 1 material and /Target/materials is its definition. In the latter case there are more than 1 materials and /Target/materials is an array containing the definitions.
The same holds for /Target/materials/composition. If the material is monoatomic, composition can be a single object defining the atomic element. Otherwise it is an array of atom definitions.