Implement the CascadesOnly simulation mode
In this case the ion_beam object selects the ion from the material present at the position where the source ion is created.
In replacement collisions, the moving ion inherits the uid of the stopped ion. Thus a I-V parir generated by a replacement sequences is still "correlated" (i.e. has the same ud)